+++ blocklets/core/api/src/hooks/pre-start.ts
@@ -11,8 +11,8 @@ if (process.env.NODE_ENV === 'development') {
     // Run database migrations
     await import('../store/migrate').then((m) => m.default());
 
-    const { runCreditMigration } = await import('./credit-migration');
-    await runCreditMigration();
+    // const { runCreditMigration } = await import('./credit-migration');
+    // await runCreditMigration();
 
     process.exit(0);
   } catch (err) {