+++ blocklets/core/src/libs/api.ts
@@ -1,11 +1,11 @@
-import axios from 'axios';
+import { createAxios } from '@blocklet/js-sdk';
 
 export const PREFIX = window.blocklet?.prefix || '/';
 
 export const API_TIMEOUT = 30 * 1000;
 export const STREAM_API_TIMEOUT = 10 * 60 * 1000;
 
-const api = axios.create({
+const api = createAxios({
   baseURL: PREFIX,
   timeout: API_TIMEOUT,
 });